// studio-web/src/api/models/createModelRequestDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: v1
 */

export interface CreateModelRequestDTO {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  configJson?: string | null;
  isPremium?: boolean;
  price?: number;
  /** @nullable */
  thumbnailFileId?: string | null;
  /** @nullable */
  modelFileId?: string | null;
}